Move `renderHtml()` into positron subfolder
